Tests/LibWeb/Text/expected/ShadowDOM/descendant-slot-invalidation-scan-stays-linear.txt

deep child color: rgb(0, 128, 0)
descendant slot scans stayed linear: true
